Can't just call removedService/addingService
The first serviceRegistration is "bound" to the serviceReference in the `addingService` method. However, if `modifiedService` were called, that serviceRegistration is "unregistered" and the new one is not bound, since the `modifiedService` method does not have a return type.

In this case we need to update the properties of the registration that change due to changes in the service reference properties. If nothing else would change we wouldn't need to do anything since the serviceReferences are mutable.

Should we needed to create a new serviceRegistration we would need to return a mutable holder and update it with the new serviceRegistration.