LPS-105151 General simplification
I think it is better to store the property map to the PID so we have the configuration available both when adding and removing. This way we can make add and remove symmetric and we can use them when creating the configuration for the first time or when we are updating it.